Fix bugs introduced in 9f6fa32a3ac65

+++ source/client/cl_input.c
@@ -211,9 +211,9 @@ static void CL_SetUcmdMovement( usercmd_t *ucmd ) {
 		CL_GameModule_AddMovement( movement );
 	}
 
-	ucmd->sidemove = bound( -127, movement[0] * 127, 127 );
-	ucmd->forwardmove = bound( -127, movement[1] * 127, 127 );
-	ucmd->upmove = bound( -127, movement[2] * 127, 127 );
+	ucmd->sidemove = bound( -127, (int)(movement[0] * 127.0f), 127 );
+	ucmd->forwardmove = bound( -127, (int)(movement[1] * 127.0f), 127 );
+	ucmd->upmove = bound( -127, (int)(movement[2] * 127.0f), 127 );
 }
 
 /*
